Add files via upload
The project revolves around the idea of better waste management and improving the quality of roads by eliminating potholes/pit from the roads with the help of an application which would be accessible to the local people. This application would let people to click pictures of the garbage collected around their home, office or any street as well as pothole/pit present anywhere on the Roads. The respective person would just have to click the picture of the heap of local garbage or pothole and upload it in on the Server with the user number and its location both in coordinates and address. The key here is to engage both the local people and working government bodies. The image which user will click and upload to the respected Firebase will also contain the exact co-ordinates of user at the time he/she uploaded the picture